created /settings GET route, finished frontend
the settings menu will be 'for show' for now, but it's created and the user settings are fetched properly. I also realized that I architected these routes as if I am using a RDBMS instead of an object store, which would've made this all much easier... but you live and you learn I guess. Next up I need to 1) add an autosave feature for the settings 2) add an updateEmail method/route 3) re-think how the mailer actually fires/monitors and 4) make the settings actually affect something